Remove legacy compose generator entirely
