+++ app/lib/metadata.py
@@ -6,14 +6,66 @@ import os
 import yaml
 import traceback
 
-from lib.composegenerator.shared.networking import assignIp, assignIpV4
-from lib.composegenerator.v2.types import Container
+from lib.citadelutils import parse_dotenv
 from lib.entropy import deriveEntropy
 from dacite import from_dict
 
 appPorts = {}
 appPortMap = {}
 disabledApps = []
+
+def assignIpV4(appId: str, containerName: str):
+    scriptDir = path.dirname(path.realpath(__file__))
+    nodeRoot = path.join(scriptDir, "..", "..", "..", "..")
+    networkingFile = path.join(nodeRoot, "apps", "networking.json")
+    envFile = path.join(nodeRoot, ".env")
+    cleanContainerName = containerName.strip()
+    # If the name still contains a newline, throw an error
+    if cleanContainerName.find("\n") != -1:
+        raise Exception("Newline in container name")
+    env_var = "APP_{}_{}_IP".format(
+        appId.upper().replace("-", "_"),
+        cleanContainerName.upper().replace("-", "_")
+    )
+    # Write a list of used IPs to the usedIpFile as JSON, and read that file to check if an IP
+    # can be used
+    usedIps = []
+    networkingData = {}
+    if path.isfile(networkingFile):
+        with open(networkingFile, 'r') as f:
+            networkingData = json.load(f)
+
+    if 'ip_addresses' in networkingData:
+        usedIps = list(networkingData['ip_addresses'].values())
+    else:
+        networkingData['ip_addresses'] = {}
+    # An IP 10.21.21.xx, with x being a random number above 40 is asigned to the container
+    # If the IP is already in use, it will be tried again until it's not in use
+    # If it's not in use, it will be added to the usedIps list and written to the usedIpFile
+    # If the usedIpsFile contains all IPs between  10.21.21.20 and  10.21.21.255 (inclusive),
+    # Throw an error, because no more IPs can be used
+    if len(usedIps) == 235:
+        raise Exception("No more IPs can be used")
+
+    if "{}-{}".format(appId, cleanContainerName) in networkingData['ip_addresses']:
+        ip = networkingData['ip_addresses']["{}-{}".format(
+            appId, cleanContainerName)]
+    else:
+        while True:
+            ip = "10.21.21." + str(random.randint(20, 255))
+            if ip not in usedIps:
+                networkingData['ip_addresses']["{}-{}".format(
+                    appId, cleanContainerName)] = ip
+                break
+
+    dotEnv = parse_dotenv(envFile)
+    if env_var in dotEnv and str(dotEnv[env_var]) == str(ip):
+        return
+
+    with open(envFile, 'a') as f:
+        f.write("{}={}\n".format(env_var, ip))
+    with open(networkingFile, 'w') as f:
+        json.dump(networkingData, f)
 
 def appPortsToMap():
     for port in appPorts:
@@ -64,9 +116,7 @@ def getAppRegistry(apps, app_path, portCache):
                         virtual_apps[implements] = []
                     virtual_apps[implements].append(app)
                 app_metadata.append(metadata)
-                if version == 2:
-                    getPortsV2App(app_yml, app)
-                elif version == 3:
+                if version == 3:
                     getPortsV3App(app_yml, app)
                 elif version == 4:
                     getPortsV4App(app_yml, app)
@@ -149,16 +199,6 @@ def validatePort(containerName, appContainer, port, appId, priority: int, isDyna
                         "priority": priority,
                         "dynamic": isDynamic,
                     }
-
-def getPortsV2App(app, appId):
-    for appContainer in app["containers"]:
-        if "port" in appContainer:
-            validatePort(appContainer["name"], appContainer, appContainer["port"], appId, 0)
-        if "ports" in appContainer:
-            for port in appContainer["ports"]:
-                realPort = int(str(port).split(":")[0])
-                validatePort(appContainer["name"], appContainer, realPort, appId, 2)
-
 
 def getPortsV3App(app, appId):
     for appContainer in app["containers"]:
