+++ app/lib/validate.py
@@ -11,27 +11,16 @@ import traceback
 scriptDir = os.path.join(os.path.dirname(os.path.realpath(__file__)), "..")
 nodeRoot = os.path.join(scriptDir, "..")
 
-with open(os.path.join(scriptDir, 'app-standard-v2.yml'), 'r') as f:
-    schemaVersion2 = yaml.safe_load(f)
-
 with open(os.path.join(nodeRoot, "db", "dependencies.yml"), "r") as file: 
   dependencies = yaml.safe_load(file)
 
 # Validates app data
 # Returns true if valid, false otherwise
 def validateApp(app: dict):
-    if 'version' in app and str(app['version']) == "2":
-        try:
-            validate(app, schemaVersion2)
-            return True
-        # Catch and log any errors, and return false
-        except Exception as e:
-            print(traceback.format_exc())
-            return False
-    elif 'version' in app and str(app['version']) == "3":
+    if 'version' in app and str(app['version']) == "3":
         # The app-cli does this validation now
         return True
-    elif 'version' not in app and 'citadel_version' not in app:
+    elif 'citadel_version' not in app:
         print("Unsupported app version")
         return False
     else:
